requirements.txt
Django
psycopg2
djangorestframework
django-filter
pytest
pytest-django
pytest-factoryboy